feat: out-of-band conversation type (#288)
Changes:
1. Defines a 'oneshot' conversation type
2. Defines a 'readd request' oneshot message, which is used to request fork recovery
3. Defines a readd intent, which is the response to the request. The creator of the intent will publish a commit and then send welcomes to the requester.

__________

Optional notes:
Readd requests need to be sent to superadmins outside of the main group (as the sender is forked), however they still need MLS level authentication and privacy guarantees.

We could create a single-use group and then send a message on it, but given only a single message is needed, an alternative is to embed the message on the immutable metadata of the group itself, and have the recipient 'receive' the message at the time of decrypting the welcome, without persisting the group (ie: use openMLS to construct the group/StagedWelcome in-memory, pull the message from the immutable metadata, then throw the group away). That would allow us to avoid:
- Delays from the sender needing to send an additional message on the group after the welcomes have been sent
- Delays from the recipient needing to sync the group after receiving the welcome to receive the message
- Any unintentional consequences around group storage, syncing on it in the future, calling maybe_update_installations, or rendering it in UI. Also means we don't create a new topic on the server.